requirements.txt
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
distro==1.9.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.30.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
neo4j==5.28.2
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.22
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
